Time: 61 ms (53.19%), Space: 43.1 MB (48.94%) - LeetHub

+++ 3553-check-if-two-chessboard-squares-have-the-same-color/3553-check-if-two-chessboard-squares-have-the-same-color.cs
@@ -1,5 +1,5 @@
 public class Solution {
     public bool CheckTwoChessboards(string coordinate1, string coordinate2) {
-        return ((int)coordinate2[0]+(int)coordinate2[1])%2==((int)coordinate1[0]+(int)coordinate1[1])%2;
+        return (coordinate2[0]+coordinate2[1])%2==(coordinate1[0]+coordinate1[1])%2;
     }
 }
